requirements.txt
aiofiles
aiohttp>=2.2.5
apscheduler
bleach
beautifulsoup4
currencyconverter
emoji
feedparser
future
jikanpy
git+https://github.com/python-telegram-bot/ptbcontrib.git
gitpython
gpytranslate
markdown2
pillow>=4.2.0
psutil
psycopg2-binary
pyrate-limiter
pyrogram
python-telegram-bot==13.9
requests
spamwatch
speedtest-cli
sqlalchemy==1.4.11
telethon==1.24.0
telegraph
wikipedia
yt-dlp
youtube-search-python
py-tgcalls==0.9.1
numpy
